Mac address deadline.

+++ resources/views/macAddresses/index.blade.php
@@ -17,19 +17,29 @@
       MAC
     </div>
 		<div class="panel-body">
+      @if($user->isStudent() && $user->group->mac_address_deadline != null)
+        Verrouillage le
+                {{ (new Carbon($user->group->mac_address_deadline))->formatLocalized('%d/%m/%Y à %H heures') }}
+                ({{ (new Carbon($user->group->mac_address_deadline))->diffForHumans()}})
+        <hr>
+      @endif
       @foreach($addresses as $address)
-        {{ $address->address }}
-        <a href="/macAddress/{{ $address->id }}/edit"><button class="btn btn-warning actionButton"><i class="fa fa-edit"></i></button></a>
-        {{Form::open(['method' => 'delete',
-                'url' => '/macAddress/'.$address->id,
-                'class' => 'deleteBtn',
-                'class' => 'form-inline',
-                'style' => 'display:inline'])}}
-          <div class="form-group">
-            <button type="submit" class="btn btn-danger actionButton"><i class="fa fa-trash"></i></button>
-          </div>
-        {{Form::close()}}
-        <HR>
+          {{ $address->address }}
+        @can('edit', $address)
+          <a href="/macAddress/{{ $address->id }}/edit"><button class="btn btn-warning actionButton"><i class="fa fa-edit"></i></button></a>
+        @endcan
+        @can('destroy', $address)
+          {{Form::open(['method' => 'delete',
+                  'url' => '/macAddress/'.$address->id,
+                  'class' => 'deleteBtn',
+                  'class' => 'form-inline',
+                  'style' => 'display:inline'])}}
+            <div class="form-group">
+              <button type="submit" class="btn btn-danger actionButton"><i class="fa fa-trash"></i></button>
+            </div>
+          {{Form::close()}}
+        @endcan
+        <hr>
       @endforeach
       @can('create', \App\Models\MacAddress::class)
         <a href="/macAddress/create"><button class="btn btn-primary actionButton"><i class="fa fa-plus"></i></button></a>
